Move composite expression tests out of test_base
